requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
google-generativeai==0.7.2
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.0.0
python-dotenv==1.0.0